requirements.txt
aenum==3.1.15
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
beanie==1.25.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Faker==24.8.0
fastapi==0.110.1
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib==1.0.4
iso8601==1.1.0
lazy-model==0.2.0
motor==3.4.0
psutil==5.9.8
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
tortoise-orm==0.20.0
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
